refactor skeletonview facade
