+++ SkeletonViewCore/Sources/API/Deprecated.swift
