--- SkeletonViewCore/Sources/API/PublicSkeletonView.swift
@@ -1,14 +1,0 @@
-//
-//  Copyright SkeletonView. All Rights Reserved.
-//
-//  Licensed under the MIT License (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//      https://opensource.org/licenses/MIT
-//
-//  PublicSkeletonView.swift
-//
-//  Created by Juanpe Catalán on 18/8/21.
-
-import Foundation
